Report on all tests that leak, not just the first
This also resets any called mocks at the end of every test. This might
change test behavior, but is required to not report on dirty mocks after
every following test.

Another option for the future might be to fingerprint the called mocks,
and only report on ones whose fingerprints have changed.